+++ src/types.ts
@@ -38,18 +38,18 @@ export enum WeightName {
 // Keys that are valid for local browser storage.
 export enum StorageKey {
     BLOCKED_ACCOUNTS = 'BlockedAccounts',
-    FAVOURITED_ACCOUNTS = WeightName.FAVOURITED_ACCOUNTS,
+    FAVOURITED_TOOTS = 'FavouritedToots',
     FEDIVERSE_TRENDING_TAGS = 'FediverseTrendingTags',
     FEDIVERSE_TRENDING_LINKS = 'FediverseTrendingLinks',
     FEDIVERSE_TRENDING_TOOTS = 'FediverseTrendingToots',
     FILTERS = 'Filters',
     FOLLOWED_ACCOUNTS = 'FollowedAccounts',
     FOLLOWED_TAGS = WeightName.FOLLOWED_TAGS,
-    HASHTAG_PARTICIPATION = WeightName.HASHTAG_PARTICIPATION,
     HOME_TIMELINE = 'HomeTimeline',
     LAST_OPENED = "LastOpened",
     MUTED_ACCOUNTS = 'MutedAccounts',
     OPENINGS = "Openings",
+    PARTICIPATED_HASHTAG_TOOTS = 'ParticipatedHashtagToots',
     POPULAR_SERVERS = 'PopularServers',
     RECENT_NOTIFICATIONS = 'RecentNotifications',
     RECENT_USER_TOOTS = 'RecentUserToots',
@@ -121,18 +121,25 @@ export type ScorerInfo = {
     scorer?: Scorer;
 };
 
+export type StorableApiObject = (
+    MastodonTag |
+    SerializableToot |
+    TrendingLink |
+    mastodon.v1.Account |
+    mastodon.v1.Notification |
+    mastodon.v1.Tag |
+    mastodon.v1.TrendLink |
+    mastodon.v2.Filter
+);
+
 // Types that are valid for browser local storage
 export type StorableObj = (
     FeedFilterSettingsSerialized |
     MastodonServersInfo |
-    MastodonTag[] |
-    SerializableToot[] |
+    StorableApiObject |
+    StorableApiObject[] |
     StringNumberDict |
-    TrendingLink[] |
     Weights |
-    mastodon.v1.Account |
-    mastodon.v1.Account[] |
-    mastodon.v2.Filter[] |
     number
 );
 
@@ -191,6 +198,15 @@ export type UserDataSerialized = {
     participatedHashtags: TagNames,
     serverSideFilters: mastodon.v2.Filter[],
 };
+
+
+// TODO: unused stuff below here
+// Check if two types are equal from: https://www.geeksforgeeks.org/how-to-test-if-two-types-are-exactly-the-same-in-typescript/
+// Usage:
+//     const areTypesEqual: EqualType<Type1, Type2> = true;
+export type EqualType<T, U> = keyof T extends keyof U
+                            ? (keyof U extends keyof T ? true : false)
+                            : false;
 
 
 // From https://dev.to/nikosanif/create-promises-with-timeout-error-in-typescript-fmm